+++ CHANGELOG.md
@@ -4,6 +4,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- Starlight charger model (naj77)
+- Metal alloyer textures (naj77)
 ### Fixed
 - Star Charged enchanted books existing (hopefully)